refactor: Update types and entities for contract and voucher
This commit refactors several parts of the codebase:

-   Updates `IInstallmentBody` to `InstallmentBody` in `contract.controller.ts` to align with the type definition.
-   Imports `EntryType` from `entry.types.ts` in `voucher.service.ts` to resolve the import issue.
-   Changes the data type of `vat_value` and `price_before_vat` from `float64` to `float` in `Contract.entity.ts` for better precision.

+++ src/entities/Contract.entity.ts
@@ -191,13 +191,13 @@ export class Contract {
   @Column('date', { nullable: true })
   property_delivery_date?: Date;
 
-  @Column('float64', { nullable: true })
+  @Column('float', { nullable: true })
   vat_value?: number;
 
   @Column('uuid', { nullable: true }) 
   vat_account_id?: string;
 
-  @Column('float64', { nullable: true })
+  @Column('float', { nullable: true })
   price_before_vat?: number;
 
   @ManyToOne(() => Tenant, { onDelete: 'CASCADE', nullable: true })
